Reorganizing some files and putting them in proper namespaces.
