+++ SimpleSmtpInterceptor.Lib/Services/AttachmentCompressor.cs
